[optimize]观看历史HistoryActivity页面和AnimeCover9类型使用Compose重写；移除NoticeActivity和WebViewActivity

+++ app/src/main/java/com/skyd/imomoe/viewmodel/HistoryViewModel.kt
@@ -6,31 +6,29 @@ import com.skyd.imomoe.appContext
 import com.skyd.imomoe.bean.HistoryBean
 import com.skyd.imomoe.database.getAppDataBase
 import com.skyd.imomoe.ext.request
-import com.skyd.imomoe.state.DataState
 import com.skyd.imomoe.util.showToast
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 
 
 class HistoryViewModel : ViewModel() {
-    val historyList: MutableStateFlow<DataState<List<Any>>> = MutableStateFlow(DataState.Empty)
-    val deleteHistory: MutableSharedFlow<HistoryBean?> = MutableSharedFlow(extraBufferCapacity = 1)
-    val deleteAllHistory: MutableSharedFlow<Boolean> = MutableSharedFlow(extraBufferCapacity = 1)
+    val uiState: MutableStateFlow<HistoryUiState> = MutableStateFlow(HistoryUiState.Refreshing())
 
     init {
         getHistoryList()
     }
 
     fun getHistoryList() {
-        historyList.tryEmit(DataState.Refreshing)
+        val currentState = uiState.value
+        val oldList = if (currentState is HistoryUiState.Success) currentState.dataList else null
+        uiState.tryEmit(HistoryUiState.Refreshing(oldList))
         request(request = { getAppDataBase().historyDao().getHistoryList() }, success = {
             it.sortWith { o1, o2 ->
                 // 负数表示按时间戳从大到小排列
                 -o1.time.compareTo(o2.time)
             }
-            historyList.tryEmit(DataState.Success(it))
+            uiState.tryEmit(HistoryUiState.Success(it))
         }, error = {
-            historyList.tryEmit(DataState.Error(it.message.orEmpty()))
+            uiState.tryEmit(HistoryUiState.Error(it.message.orEmpty()))
             "${appContext.getString(R.string.get_data_failed)}\n${it.message}".showToast()
         })
     }
@@ -39,23 +37,33 @@ class HistoryViewModel : ViewModel() {
         request(request = {
             getAppDataBase().historyDao().deleteHistory(historyBean.animeUrl)
             getHistoryList()
-        }, success = {
-            deleteHistory.tryEmit(historyBean)
         }, error = {
-            deleteHistory.tryEmit(null)
             "${appContext.getString(R.string.delete_failed)}\n${it.message}".showToast()
         })
     }
 
     fun deleteAllHistory() {
+        val currentState = uiState.value
+        if (currentState is HistoryUiState.Success && currentState.dataList.isEmpty()) return
         request(request = {
             getAppDataBase().historyDao().deleteAllHistory()
             getHistoryList()
-        }, success = {
-            deleteAllHistory.tryEmit(true)
         }, error = {
-            deleteAllHistory.tryEmit(false)
             "${appContext.getString(R.string.delete_failed)}\n${it.message}".showToast()
         })
+    }
+}
+
+sealed interface HistoryUiState {
+    data class Success(val dataList: List<Any>) : HistoryUiState
+    data class Error(val message: String = "") : HistoryUiState
+    data class Refreshing(val oldList: List<Any>? = null) : HistoryUiState
+
+    fun readOrNull(): List<Any>? {
+        return when (this) {
+            is Success -> dataList
+            is Refreshing -> oldList
+            else -> null
+        }
     }
 }
